[I] Refactor BaseDropdown to Dropdown
