+++ components/atomic/dropdown/Dropdown/Dropdown.stories.tsx
@@ -1,8 +1,7 @@
-import Dropdown from "./Dropdown";
+import Dropdown, { BaseDisclosureProps } from "./Dropdown";
 import { Story } from "@storybook/react";
 import { ButtonControl } from "components/atomic/buttons";
 import { IconFactory } from "components/factories";
-import { BaseDisclosureProps } from "./BaseDropdown";
 
 type Props = React.ComponentProps<typeof Dropdown>;
